requirements.txt
boto3>=1.37.0
botocore>=1.37.0
pyttsx3>=2.90
colorama>=0.4.0
anthropic>=0.7.0
requests>=2.25.0
google-generativeai>=0.7.0
openai>=1.0.0
google-genai>=1.54.0